split into SPA and implement profile page
